Model material overrides * Made template naming in Event.hpp and ValueRange.hpp more standardized. * Warning Level 3 no longer generates any warnings (other than third-party code) * Simplified the `EngineObject` class: It is no longer a "singleton" (i.e. No `::Get` and no `::Singleton` members) * The model loading system now appends "/scene.gltf" to the end of model paths * Models can specify their own Material by having a "material.mtl" in their folder in `resources/models` * Specular exponents and Multipliers defined in the Material now work again (but they seem to be way too strong) * SDL and ImGui have been updated * Logs are now saved every second in addition to being saved on shutdown * `FileRW` now checks if the file handle could be opened in `::WriteFile` * Updated the comment block in `Main.cpp` to add the new controls and a link to the GitHub * Removed redundant `#include`s * The Engine is now stack-allocated * The result of `IMGUI_CHECKVERSION` is no longer discarded (another class Mr Victor Gordan moment), but it's still just compiled in with everything else so there's no way for the version mismatch check to ever trigger * Changed some error messages and crash message formatting * ImGui and SDL are explicitly shutdown now

+++ include/gameobject/Base3D.hpp
@@ -25,8 +25,8 @@ public:
 	Vector3 AngularVelocity;
 	double Mass = 1.0f;
 
-	double Transparency = 0.0f;
-	double Reflectivity = 0.0f;
+	float Transparency = 0.0f;
+	float Reflectivity = 0.0f;
 
 	Color ColorRGB = Color(1.0f, 1.0f, 1.0f);
